+++ tests/Socialite/GoogleProviderIdTokenTest.php
@@ -13,6 +13,7 @@ use Hypervel\Socialite\Two\Exceptions\InvalidAudienceException;
 use Hypervel\Socialite\Two\Exceptions\InvalidIssuerException;
 use Hypervel\Socialite\Two\GoogleProvider;
 use Hypervel\Socialite\Two\User;
+use Hypervel\Tests\Socialite\Fixtures\CreatesJwksFixtures;
 use Hypervel\Tests\TestCase;
 use Mockery as m;
 use PHPUnit\Framework\Attributes\DataProvider;
@@ -22,6 +23,8 @@ use ReflectionMethod;
 
 class GoogleProviderIdTokenTest extends TestCase
 {
+    use CreatesJwksFixtures;
+
     public function testItCanDetectJwtTokens(): void
     {
         $provider = $this->getProvider();
@@ -226,43 +229,5 @@ class GoogleProviderIdTokenTest extends TestCase
             'iat' => time(),
             'exp' => time() + 3600,
         ], $key['private'], 'RS256', $key['kid']);
-    }
-
-    private function createRsaKeyPair(string $kid): array
-    {
-        $key = openssl_pkey_new([
-            'private_key_bits' => 2048,
-            'private_key_type' => OPENSSL_KEYTYPE_RSA,
-        ]);
-
-        if ($key === false) {
-            $this->fail('Unable to generate RSA key pair for Google ID token test.');
-        }
-
-        openssl_pkey_export($key, $privateKey);
-        $details = openssl_pkey_get_details($key);
-
-        return [
-            'kid' => $kid,
-            'private' => $privateKey,
-            'jwk' => [
-                'kid' => $kid,
-                'kty' => 'RSA',
-                'use' => 'sig',
-                'alg' => 'RS256',
-                'n' => $this->base64UrlEncode($details['rsa']['n']),
-                'e' => $this->base64UrlEncode($details['rsa']['e']),
-            ],
-        ];
-    }
-
-    private function jwks(array $key): array
-    {
-        return ['keys' => [$key['jwk']]];
-    }
-
-    private function base64UrlEncode(string $value): string
-    {
-        return rtrim(strtr(base64_encode($value), '+/', '-_'), '=');
     }
 }
